# tipu-api/openapi.yaml
openapi: 3.0.0
info:
  title: TIPU Academy API
  version: 1.0.0
  description: Online tutoring platform API with Firebase and Stripe integration
  contact:
    name: TIPU Academy Team
    email: support@tipuacademy.com
  license:
    name: MIT
servers:
  - url: http://localhost:8888
    description: Development server
  - url: https://api.tipuacademy.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT authentication token
  schemas:
    User:
      type: object
      required:
        - uid
        - email
        - displayName
        - role
      properties:
        uid:
          type: string
          description: Firebase Auth user ID
          example: abc123xyz789
        email:
          type: string
          format: email
          description: User email address
          example: student@example.com
        displayName:
          type: string
          description: User full name
          minLength: 2
          maxLength: 100
          example: John Smith
        role:
          type: string
          enum:
            - student
            - tutor
            - parent
            - admin
          description: User role in the system
          example: student
        photoURL:
          type: string
          format: uri
          description: Profile photo URL
          example: https://example.com/photo.jpg
        dateOfBirth:
          type: string
          format: date-time
          description: Date of birth (for 18+ chat eligibility)
          example: "2005-06-15T00:00:00Z"
        parentId:
          type: string
          description: Parent user ID (for students under 18)
          example: parent123
        subjects:
          type: array
          items:
            type: string
            enum:
              - Maths
              - Physics
              - Computer Science
              - Python
          description: Subjects (for tutors)
          example:
            - Maths
            - Physics
        bio:
          type: string
          description: Tutor biography
          maxLength: 1000
          example: Experienced Maths tutor with 10+ years...
        isApproved:
          type: boolean
          description: Admin approval status (for tutors)
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-01-20T14:20:00Z"
...
    Payment:
      type: object
      properties:
        id:
          type: string
          description: Payment ID
          example: pay123
        bookingId:
          type: string
          description: Associated booking ID
          example: booking123
        amount:
          type: number
          description: Amount in pence
          minimum: 0
          example: 4500
        currency:
          type: string
          description: Currency code
          default: gbp
          example: gbp
        status:
          type: string
          enum:
            - pending
            - succeeded
            - failed
            - refunded
          description: Payment status
          example: succeeded
        paymentIntentId:
          type: string
          description: Stripe Payment Intent ID
          example: pi_123456789
        createdAt:
          type: string
          format: date-time
          example: "2025-11-23T10:00:00Z"
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Resource not found
        message:
          type: string
          description: Detailed error message
          example: The requested booking does not exist
        code:
          type: string
          description: Error code
          example: BOOKING_NOT_FOUND
paths:
...
  /v1/payments/create-intent:
    post:
      tags:
        - Payments
      summary: Create Stripe Payment Intent
      description: Create a payment intent for a booking. Returns client secret for Stripe.js on frontend.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookingId
                - amount
              properties:
                bookingId:
                  type: string
                  description: ID of the booking to pay for
                  example: booking123
                amount:
                  type: number
                  description: Amount in pence (£45.00 = 4500)
                  minimum: 100
                  example: 4500
                currency:
                  type: string
                  description: Currency code
                  default: gbp
                  enum:
                    - gbp
                  example: gbp
      responses:
        "200":
          description: Payment intent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientSecret:
                    type: string
                    description: Stripe client secret for frontend
                    example: pi_123_secret_456
                  paymentIntentId:
                    type: string
                    description: Stripe Payment Intent ID
                    example: pi_123456789
        "400":
          description: Invalid booking ID or amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Booking not found
  /v1/payments/history:
    get:
      tags:
        - Payments
      summary: Get payment history
      description: Retrieve payment history for the authenticated user (student/parent)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of payments to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of payments to skip (for pagination)
      responses:
        "200":
          description: Payment history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Payment"
              example:
                payments:
                  - id: pay123
                    bookingId: booking123
                    amount: 4500
                    currency: gbp
                    status: succeeded
                    paymentIntentId: pi_123456789
                    createdAt: "2025-11-23T10:00:00Z"
        "401":
          description: Unauthorized
  /v1/payments/connect-account:
    post:
      tags:
        - Payments
      summary: Create Stripe Connect account for tutor
      description: Create a Stripe Connect account to receive payouts. Returns onboarding URL for tutor to complete setup.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connect account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId:
                    type: string
                    description: Stripe Connect account ID
                    example: acct_123456789
                  onboardingUrl:
                    type: string
                    format: uri
                    description: URL for tutor to complete Stripe onboarding
                    example: https://connect.stripe.com/setup/e/acct_123/xyz
        "400":
          description: Account already exists or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not a tutor
  /v1/payments/webhook:
    post:
      tags:
        - Payments
      summary: Stripe webhook endpoint
      description: |
        Handle Stripe webhook events. This endpoint is called by Stripe to notify about payment events.

        Handled events:
        - payment_intent.succeeded - Mark booking as paid
        - payment_intent.payment_failed - Log payment failure

        **Note:** This endpoint requires Stripe signature verification and should only be called by Stripe.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook event payload
      responses:
        "200":
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
        "400":
          description: Webhook signature verification failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Webhook Error: Invalid signature"
        "500":
          description: Error handling webhook event
  /v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user profile by ID
      description: Retrieve detailed user profile information
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
          example: user123
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags:
        - Users
      summary: Update user profile
      description: Update user profile information. Users can only update their own profile unless they are an admin.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID to update
          example: user123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: John Smith
                photoURL:
                  type: string
                  format: uri
                  example: https://example.com/photo.jpg
                bio:
                  type: string
                  maxLength: 1000
                  description: Tutor biography (tutors only)
                  example: Experienced Maths tutor with 10+ years of experience
                subjects:
                  type: array
                  items:
                    type: string
                    enum:
                      - Maths
                      - Physics
                      - Computer Science
                      - Python
                  description: Subjects taught (tutors only)
                  example:
                    - Maths
                    - Physics
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Cannot update another user's profile
        "404":
          description: User not found
  /v1/users/tutors/all:
    get:
      tags:
        - Users
      summary: Get all approved tutors
      description: Retrieve a list of all approved and active tutors
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of tutors to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of tutors to skip (for pagination)
      responses:
        "200":
          description: Tutors retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tutors:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                tutors:
                  - uid: tutor123
                    email: tutor@example.com
                    displayName: Jane Doe
                    role: tutor
                    subjects:
                      - Maths
                      - Physics
                    bio: Experienced tutor with 10+ years
                    isApproved: true
        "401":
          description: Unauthorized
  /v1/users/tutors/subject/{subject}:
    get:
      tags:
        - Users
      summary: Get tutors by subject
      description: Retrieve approved tutors who teach a specific subject
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject
          required: true
          schema:
            type: string
            enum:
              - Maths
              - Physics
              - Computer Science
              - Python
          description: Subject to filter by
          example: Maths
      responses:
        "200":
          description: Tutors retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tutors:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                tutors:
                  - uid: tutor123
                    email: tutor@example.com
                    displayName: Jane Doe
                    role: tutor
                    subjects:
                      - Maths
                      - Physics
                    bio: Experienced Maths tutor
                    isApproved: true
        "400":
          description: Invalid subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: No tutors found for this subject
...